requirements.txt
Pillow==9.1.0
requests~=2.25.1

lxml~=4.9.1
aiohttp~=3.8.3